refactor(report): [chart] extract logic for chart options
