+++ libs/bublik/features/run-report/src/lib/run-report-chart/run-report-chart.component.tsx
@@ -1,20 +1,19 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 /* SPDX-FileCopyrightText: 2024 OKTET LTD */
-import { ComponentProps, useEffect, useMemo, useRef, useState } from 'react';
-import { z } from 'zod';
+import { useEffect, useRef, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { toast } from 'sonner';
 import ReactEChartsCore from 'echarts-for-react/lib/core';
 
-import { Plot, chartStyles } from '@/shared/charts';
+import { Plot } from '@/shared/charts';
 import { ReportChart } from '@/shared/types';
 import { LogPreviewContainer } from '@/bublik/features/log-preview-drawer';
 import { usePlatformSpecificCtrl } from '@/shared/hooks';
 
-const ParamsSchema = z.object({
-	seriesIndex: z.number(),
-	dataIndex: z.number()
-});
+import {
+	ParamsSchema,
+	resolveRunReportChartOptions
+} from './run-report-chart.utils';
 
 interface RunReportChartProps {
 	chart: ReportChart;
@@ -25,57 +24,6 @@ function RunReportChart(props: RunReportChartProps) {
 	const { runId } = useParams<{ runId: string }>();
 	const [resultId, setResultId] = useState<number>();
 	const [open, setOpen] = useState(false);
-
-	const series = useMemo<
-		ComponentProps<typeof Plot>['options']['series']
-	>(() => {
-		return chart.data.map((d) => ({
-			name:
-				chart.series_label === null && chart.data.length === 1
-					? chart.axis_y.label
-					: d.series,
-			type: 'line',
-			data: d.points.map((d) => [d?.[chart.axis_x.key], d?.[chart.axis_y.key]]),
-			itemStyle: {
-				color: (params: unknown) => {
-					const parsedParams = ParamsSchema.safeParse(params);
-					if (!parsedParams.success) return '#65cd84';
-
-					const point = d.points[parsedParams.data.dataIndex];
-					const hasError = point?.metadata?.has_error;
-
-					if (hasError) return '#f95c78';
-					return '#65cd84';
-				}
-			},
-			symbol: (_, params: unknown) => {
-				const parsedParams = ParamsSchema.safeParse(params);
-				if (!parsedParams.success) return 'emptyCircle';
-
-				const point = d.points[parsedParams.data.dataIndex];
-				const hasError = point?.metadata?.has_error;
-
-				if (hasError) return 'diamond';
-				return 'emptyCircle';
-			},
-			symbolSize: (_, params: unknown) => {
-				const parsedParams = ParamsSchema.safeParse(params);
-				if (!parsedParams.success) return 8;
-
-				const point = d.points[parsedParams.data.dataIndex];
-				const hasError = point?.metadata?.has_error;
-
-				if (hasError) return 16;
-				return 8;
-			}
-		}));
-	}, [
-		chart.axis_x.key,
-		chart.axis_y.key,
-		chart.axis_y.label,
-		chart.data,
-		chart.series_label
-	]);
 
 	const chartRef = useRef<ReactEChartsCore>(null);
 
@@ -110,9 +58,6 @@ function RunReportChart(props: RunReportChartProps) {
 		};
 	}, [chart.data]);
 
-	const isPressed = usePlatformSpecificCtrl();
-	const dataZoom = [{}, { type: 'inside', zoomLock: !isPressed }];
-
 	useEffect(() => {
 		function handleResize() {
 			chartRef.current?.getEchartsInstance().resize();
@@ -123,6 +68,10 @@ function RunReportChart(props: RunReportChartProps) {
 			window.removeEventListener('resize', handleResize);
 		};
 	}, []);
+
+	const isCtrlPressed = usePlatformSpecificCtrl();
+
+	const options = resolveRunReportChartOptions({ chart, isCtrlPressed });
 
 	return (
 		<>
@@ -137,55 +86,7 @@ function RunReportChart(props: RunReportChartProps) {
 				<Plot
 					ref={chartRef}
 					notMerge={false}
-					options={{
-						toolbox: { top: 9999, feature: { dataZoom: {} } },
-						title: {
-							show: Boolean(chart.series_label),
-							text: chart.series_label,
-							left: 'center',
-							textStyle: { fontSize: 12, fontFamily: 'Inter' }
-						},
-						tooltip: {
-							trigger: 'axis',
-							textStyle: chartStyles.text,
-							extraCssText: 'shadow-popover rounded-lg',
-							axisPointer: { type: 'shadow' }
-						},
-						legend: {
-							data: chart.data.map((s) => s.series ?? ''),
-							show: Boolean(chart.series_label),
-							padding: chart.series_label ? [25, 0, 0, 0] : undefined
-						},
-						grid: {
-							containLabel: true,
-							top:
-								Boolean(chart.series_label) && chart.data.length >= 4
-									? '28%'
-									: '15%',
-							right: '7%',
-							left: '5%',
-							bottom: '20%'
-						},
-						dataZoom,
-						xAxis: {
-							type: 'value',
-							name: chart.axis_x.label,
-							nameLocation: 'middle',
-							nameGap: 30,
-							nameTextStyle: chartStyles.text,
-							axisLabel: { ...chartStyles.text },
-							scale: true
-						},
-						yAxis: {
-							type: 'value',
-							name: chart.axis_y.label,
-							nameGap: 20,
-							nameLocation: 'end',
-							axisLabel: chartStyles.text,
-							nameTextStyle: { ...chartStyles.text, align: 'left' }
-						},
-						series: series
-					}}
+					options={options}
 					style={{ height: '100%' }}
 				/>
 			</div>
